updated map and dropdown zone/county selection
